Allow to register OIDC tenants programmatically
