+++ extensions/oidc/runtime/src/main/java/io/quarkus/oidc/runtime/TenantConfigBean.java
@@ -8,28 +8,26 @@ import jakarta.enterprise.context.spi.CreationalContext;
 
 import io.quarkus.arc.BeanDestroyer;
 import io.quarkus.oidc.OidcTenantConfig;
+import io.quarkus.tls.TlsConfigurationRegistry;
 import io.smallrye.mutiny.Uni;
+import io.vertx.core.Vertx;
 
-public class TenantConfigBean {
+public final class TenantConfigBean {
 
     private final Map<String, TenantConfigContext> staticTenantsConfig;
     private final Map<String, TenantConfigContext> dynamicTenantsConfig;
     private final TenantConfigContext defaultTenant;
     private final TenantContextFactory tenantContextFactory;
 
-    @FunctionalInterface
-    public interface TenantContextFactory {
-        Uni<TenantConfigContext> create(OidcTenantConfig oidcTenantConfig);
-    }
-
-    public TenantConfigBean(
-            Map<String, TenantConfigContext> staticTenantsConfig,
-            TenantConfigContext defaultTenant,
-            TenantContextFactory tenantContextFactory) {
-        this.staticTenantsConfig = Map.copyOf(staticTenantsConfig);
+    TenantConfigBean(Vertx vertx, TlsConfigurationRegistry tlsConfigurationRegistry, OidcImpl oidc,
+            boolean securityEventsEnabled) {
+        this.tenantContextFactory = new TenantContextFactory(vertx, tlsConfigurationRegistry, securityEventsEnabled);
         this.dynamicTenantsConfig = new ConcurrentHashMap<>();
-        this.defaultTenant = defaultTenant;
-        this.tenantContextFactory = tenantContextFactory;
+
+        this.staticTenantsConfig = tenantContextFactory.createStaticTenantConfigs(oidc.getStaticTenantConfigs(),
+                oidc.getDefaultTenantConfig());
+        this.defaultTenant = tenantContextFactory.createDefaultTenantConfig(oidc.getStaticTenantConfigs(),
+                oidc.getDefaultTenantConfig());
     }
 
     public Uni<TenantConfigContext> createDynamicTenantContext(OidcTenantConfig oidcConfig) {
@@ -40,7 +38,7 @@ public class TenantConfigBean {
             return Uni.createFrom().item(tenant);
         }
 
-        return tenantContextFactory.create(oidcConfig).onItem().transform(
+        return tenantContextFactory.createDynamic(oidcConfig).onItem().transform(
                 new Function<TenantConfigContext, TenantConfigContext>() {
                     @Override
                     public TenantConfigContext apply(TenantConfigContext t) {
